requirements.txt
asgiref==3.7.2
certifi==2024.2.2
Django==5.0.1
django-elasticsearch-dsl==8.0
djangorestframework==3.14.0
elastic-transport==8.12.0
elasticsearch==8.12.0
elasticsearch-dsl==8.12.0
psycopg==3.1.17
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0